+++ packages/pak-files-parser/src/app/generators/npcs/known-portrait-configurations.const.ts
@@ -1,0 +1,46 @@
+export const knownEmotions = [
+    'Angry',
+    'Annoyed',
+    'Bloated',
+    'Blushing',
+    'Concerned',
+    'Daydreaming',
+    'Deafult',
+    'Default',
+    'Determined',
+    'Doubt',
+    'Happy',
+    'Happy02',
+    'Happy2',
+    'Laugh',
+    'Panic',
+    'Sad',
+    'Shocked',
+    'Smile',
+    'Smirk',
+    'Suprised',
+    'Surprised',
+    'Surprised02',
+    'Thinking',
+    'Unimpressed',
+    'blushing',
+    'happy',
+
+].map(s => s.toLowerCase());
+
+export const knownAppearances = [
+    'BS',
+    'BathingSuit',
+    'Fall',
+    'Night',
+    'Spring',
+    'Summer',
+    'Wedding',
+    'Winter',
+    'Yoga',
+].map(s => s.toLowerCase());
+
+export const knownSpecialForms = [
+    'RecCenter',
+    'Human',
+].map(s => s.toLowerCase())